Reorganising namespaces.
