+++ NoteTaker/Maui.NoteTaker.Sample/Maui.NoteTaker.Sample/AppShell.xaml
@@ -3,18 +3,18 @@
     x:Class="Maui.NoteTaker.Sample.AppShell"
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    xmlns:local="clr-namespace:Maui.NoteTaker.Sample"
+    xmlns:views="clr-namespace:Maui.NoteTaker.Sample.Views"
     Shell.FlyoutBehavior="Disabled">
 
     <TabBar>
         <ShellContent
             Title="Notes"
-            ContentTemplate="{DataTemplate local:NotePage}"
+            ContentTemplate="{DataTemplate views:NotePage}"
             Icon="icon_notes" />
 
         <ShellContent
             Title="About"
-            ContentTemplate="{DataTemplate local:AboutPage}"
+            ContentTemplate="{DataTemplate views:AboutPage}"
             Icon="icon_about" />
     </TabBar>
