+++ NoteTaker/Maui.NoteTaker.Sample/Maui.NoteTaker.Sample/Views/AboutPage.xaml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="Maui.NoteTaker.Sample.AboutPage">
+             x:Class="Maui.NoteTaker.Sample.Views.AboutPage">
     <VerticalStackLayout Spacing="10" Margin="10">
         <HorizontalStackLayout Spacing="10">
             <Image Source="dotnet_bot.png"
